Add shared Mac-style audio gate indicator
