+++ static/js/theKing.js
@@ -1,3 +1,5 @@
+import { initializeTheKingVideo } from './modules/theKingVideo.js';
+
 const macHdWindow = document.getElementById('mac-hd-window');
 const theKingWindow = document.getElementById('the-king-window');
 const theKingVideo = document.getElementById('the-king-video');
@@ -113,315 +115,11 @@ if (theKingWindow) {
   setupDraggable(theKingWindow);
 }
 
-function initializeTheKingVideo(video) {
-  if (!video) {
-    return;
-  }
-
-  const activationEvents = ['pointerdown', 'touchstart', 'keydown'];
-  const activationHandlers = new Map();
-  let activationCallback = null;
-  let awaitingAudioUnlock = false;
-  let pendingPlay = null;
-  let audioRetryTimer = null;
-  let audioConsent = false;
-
-  const AUDIO_CONSENT_STORAGE_KEY = 'theKing:audio-consent';
-
-  try {
-    audioConsent =
-      window.localStorage.getItem(AUDIO_CONSENT_STORAGE_KEY) === 'granted';
-  } catch (error) {
-    audioConsent = false;
-  }
-
-  const persistAudioConsent = granted => {
-    audioConsent = granted;
-    try {
-      if (granted) {
-        window.localStorage.setItem(AUDIO_CONSENT_STORAGE_KEY, 'granted');
-      } else {
-        window.localStorage.removeItem(AUDIO_CONSENT_STORAGE_KEY);
-      }
-    } catch (error) {
-      // Ignore storage errors (private mode, etc.).
-    }
-  };
-
-  const audioGate = (() => {
-    if (!theKingWindow) {
-      return null;
-    }
-    const gate = document.createElement('div');
-    gate.id = 'the-king-audio-gate';
-    gate.setAttribute('role', 'status');
-    gate.setAttribute('aria-live', 'polite');
-    gate.hidden = true;
-    gate.textContent = 'tap, click, or press any key to enable sound';
-    theKingWindow.appendChild(gate);
-    return gate;
-  })();
-
-  const showAudioGate = () => {
-    if (audioGate) {
-      audioGate.hidden = false;
-    }
-  };
-
-  const hideAudioGate = () => {
-    if (audioGate) {
-      audioGate.hidden = true;
-    }
-  };
-
-  const ensureAttributes = () => {
-    video.autoplay = true;
-    video.loop = true;
-    video.playsInline = true;
-    video.setAttribute('autoplay', '');
-    video.setAttribute('loop', '');
-    video.setAttribute('playsinline', '');
-    video.setAttribute('webkit-playsinline', '');
-  };
-
-  const configureMuted = () => {
-    video.defaultMuted = true;
-    video.muted = true;
-    if (!video.hasAttribute('muted')) {
-      video.setAttribute('muted', '');
-    }
-  };
-
-  const configureUnmuted = () => {
-    video.defaultMuted = false;
-    video.muted = false;
-    if (video.hasAttribute('muted')) {
-      video.removeAttribute('muted');
-    }
-    if (typeof video.volume === 'number' && video.volume < 1) {
-      try {
-        video.volume = 1;
-      } catch (error) {
-        // Ignore platforms that disallow programmatic volume changes.
-      }
-    }
-  };
-
-  const clearActivationHandlers = () => {
-    activationHandlers.forEach((handler, eventName) => {
-      window.removeEventListener(eventName, handler, { capture: true });
-    });
-    activationHandlers.clear();
-    activationCallback = null;
-  };
-
-  const clearAudioRetryTimer = () => {
-    if (audioRetryTimer !== null) {
-      window.clearTimeout(audioRetryTimer);
-      audioRetryTimer = null;
-    }
-  };
-
-  const requestActivation = callback => {
-    if (activationCallback === callback) {
-      return;
-    }
-    clearActivationHandlers();
-    activationCallback = callback;
-    activationEvents.forEach(eventName => {
-      const handler = () => {
-        clearActivationHandlers();
-        callback();
-      };
-      activationHandlers.set(eventName, handler);
-      window.addEventListener(eventName, handler, {
-        capture: true,
-        once: true,
-        passive: true,
-      });
-    });
-  };
-
-  const handleAudioUnlocked = () => {
-    awaitingAudioUnlock = false;
-    hideAudioGate();
-    clearAudioRetryTimer();
-    clearActivationHandlers();
-    persistAudioConsent(true);
-  };
-
-  const activationPlay = () => {
-    playVideo({
-      mutedFallbackAllowed: false,
-      startMuted: false,
-    }).catch(error => {
-      if (error && error.name !== 'NotAllowedError') {
-        scheduleAudioRecovery();
-      }
-    });
-  };
-
-  const beginAwaitingAudioUnlock = () => {
-    if (!awaitingAudioUnlock) {
-      awaitingAudioUnlock = true;
-      scheduleAudioRecovery();
-    }
-    showAudioGate();
-    requestActivation(activationPlay);
-  };
-
-  const scheduleAudioRecovery = () => {
-    if (!awaitingAudioUnlock) {
-      return;
-    }
-    if (audioRetryTimer !== null) {
-      return;
-    }
-    audioRetryTimer = window.setTimeout(() => {
-      audioRetryTimer = null;
-      if (!awaitingAudioUnlock) {
-        return;
-      }
-      playVideo({
-        mutedFallbackAllowed: false,
-        startMuted: false,
-      }).catch(error => {
-        if (error && error.name !== 'NotAllowedError') {
-          scheduleAudioRecovery();
-        }
-      });
-    }, 250);
-  };
-
-  const playVideoInternal = async ({
-    mutedFallbackAllowed,
-    startMuted,
-  }) => {
-    ensureAttributes();
-
-    if (startMuted) {
-      configureMuted();
-    } else {
-      configureUnmuted();
-    }
-
-    try {
-      await video.play();
-    } catch (error) {
-      if (error && error.name === 'NotAllowedError') {
-        if (!startMuted && mutedFallbackAllowed) {
-          await playVideoInternal({
-            mutedFallbackAllowed: false,
-            startMuted: true,
-          });
-          return;
-        }
-
-        configureMuted();
-        beginAwaitingAudioUnlock();
-      }
-      throw error;
-    }
-
-    if (video.muted) {
-      beginAwaitingAudioUnlock();
-    } else {
-      handleAudioUnlocked();
-    }
-  };
-
-  const playVideo = (options = {}) => {
-    if (pendingPlay) {
-      return pendingPlay;
-    }
-
-    const mergedOptions = {
-      mutedFallbackAllowed: true,
-      startMuted: awaitingAudioUnlock || !audioConsent,
-      ...options,
-    };
-
-    pendingPlay = playVideoInternal(mergedOptions)
-      .catch(error => {
-        throw error;
-      })
-      .finally(() => {
-        pendingPlay = null;
-      });
-
-    return pendingPlay;
-  };
-
-  const resumePlayback = () => {
-    playVideo().catch(() => {});
-  };
-
-  const handleVisibilityChange = () => {
-    if (document.visibilityState === 'visible') {
-      resumePlayback();
-      if (awaitingAudioUnlock) {
-        scheduleAudioRecovery();
-      }
-    }
-  };
-
-  const handlePause = () => {
-    if (video.ended) {
-      return;
-    }
-    resumePlayback();
-  };
-
-  const handleEnded = () => {
-    if (!video.loop) {
-      video.currentTime = 0;
-    }
-    resumePlayback();
-  };
-
-  const handleStall = () => {
-    if (video.paused || video.readyState < HTMLMediaElement.HAVE_CURRENT_DATA) {
-      resumePlayback();
-    }
-  };
-
-  document.addEventListener('visibilitychange', handleVisibilityChange);
-  video.addEventListener('pause', handlePause);
-  video.addEventListener('ended', handleEnded);
-  video.addEventListener('stalled', handleStall);
-  video.addEventListener('suspend', handleStall);
-  video.addEventListener('waiting', handleStall);
-  video.addEventListener('emptied', handleStall);
-  video.addEventListener('playing', () => {
-    if (awaitingAudioUnlock) {
-      scheduleAudioRecovery();
-    }
-  });
-  video.addEventListener('timeupdate', () => {
-    if (awaitingAudioUnlock) {
-      scheduleAudioRecovery();
-    }
-  });
-  video.addEventListener('loadeddata', () => {
-    if (awaitingAudioUnlock) {
-      scheduleAudioRecovery();
-    }
-  });
-  video.addEventListener('volumechange', () => {
-    if (!video.muted) {
-      handleAudioUnlocked();
-    } else if (awaitingAudioUnlock) {
-      showAudioGate();
-    }
-  });
-
-  ensureAttributes();
-  resumePlayback();
-}
-
-if (theKingVideo) {
-  initializeTheKingVideo(theKingVideo);
-}
+const theKingAudioController = theKingVideo
+  ? initializeTheKingVideo(theKingVideo, {
+      gateContainer: theKingWindow,
+    })
+  : null;
 
 window.setTimeout(() => {
   if (macHdWindow) {
@@ -431,6 +129,7 @@ window.setTimeout(() => {
     theKingWindow.style.display = 'block';
     bringToFront(theKingWindow);
   }
+  theKingAudioController?.resume();
 }, 2500);
 
 if (appleDesktop && theKingBlur) {
